Fix bug when only one $ or &, and restruct files
