+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/20 00:02:21 by migugar2          #+#    #+#             */
-/*   Updated: 2025/07/24 20:10:43 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/07/24 22:26:01 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,8 +68,8 @@ typedef enum e_lxstate
 	LX_IN_DOUBLE_Q,
 	LX_PARAM,
 	LX_EOF,
-	LX_DIE, // Unvalid state, without error
-	LX_ERR // Malloc error
+	LX_DIE,
+	LX_ERR
 }	t_lxstate;
 
 typedef struct s_slice
@@ -107,6 +107,25 @@ typedef struct s_lexer
 
 typedef t_lxstate	(*t_lxhandler)(t_lexer *lx);
 
-int	tokenize(const char *input, t_tok **out);
+t_seg		*new_seg(t_segtype k, const char *ptr, size_t len);
+int			add_seg(t_lexer *lx, t_segtype type, const char *ptr, size_t len);
+
+t_tok		*new_tok(t_toktype type);
+int			start_word(t_lexer *lx);
+void		tok_push(t_tok **head, t_tok **tail, t_tok *tok);
+void		emit_tok(t_lexer *lx);
+t_lxstate	emit_op(t_lexer *lx, t_toktype type, size_t len);
+
+void		lx_advance_n(t_lexer *lx, size_t n);
+void		lx_advance(t_lexer *lx);
+
+t_lxstate	handle_in_single_q(t_lexer *lx);
+t_lxstate	handle_in_double_q(t_lexer *lx);
+t_lxstate	handle_param(t_lexer *lx);
+t_lxstate	handle_finish(t_lexer *lx);
+
+t_lxstate	handle_general(t_lexer *lx);
+
+int			tokenize(const char *input, t_tok **out);
 
 #endif
