+++ source/lexer/general.c
@@ -1,0 +1,94 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   general.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/24 21:09:56 by migugar2          #+#    #+#             */
+/*   Updated: 2025/07/24 22:28:52 by migugar2         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+static t_lxstate	handle_general_word(t_lexer *lx)
+{
+	const char	*start;
+
+	start = lx->cur;
+	if (start_word(lx) == 1)
+		return (LX_ERR);
+	while (*lx->cur)
+	{
+		if (ft_isspace(*lx->cur)
+			|| *lx->cur == '|' || *lx->cur == '<' || *lx->cur == '>'
+			|| *lx->cur == '*' || *lx->cur == '(' || *lx->cur == ')'
+			|| *lx->cur == '\'' || *lx->cur == '"'
+			|| (*lx->cur == '&' && lx->cur[1] == '&')
+			|| (*lx->cur == '$' && (lx->cur[1] == '?'
+					|| lx->cur[1] == '_' || ft_isalnum(lx->cur[1]))))
+			break ;
+		lx_advance(lx);
+	}
+	if (add_seg(lx, SEG_TEXT, start, lx->cur - start) == 1)
+		return (ft_free((void **)&lx->tok), LX_ERR); // TODO: Free functions
+	return (LX_GENERAL);
+}
+
+static t_lxstate	handle_general_2(t_lexer *lx)
+{
+	if (*lx->cur == ')')
+		return (emit_op(lx, T_RPAREN, 1));
+	if (*lx->cur == '\'')
+		return (lx_advance(lx), LX_IN_SINGLE_Q);
+	if (*lx->cur == '"')
+		return (lx_advance(lx), LX_IN_DOUBLE_Q);
+	if (*lx->cur == '$'
+		&& (lx->cur[1] == '?' || lx->cur[1] == '_' || ft_isalnum(lx->cur[1])))
+	{
+		if (start_word(lx) == 1)
+			return (LX_ERR);
+		lx_advance(lx);
+		return (LX_PARAM);
+	}
+	if (*lx->cur == '*')
+	{
+		if (start_word(lx) == 1)
+			return (LX_ERR);
+		if (add_seg(lx, SEG_WILDCARD, lx->cur, 1) == 1)
+			return (ft_free((void **)&lx->tok), LX_ERR);
+		lx_advance(lx);
+		return (LX_GENERAL);
+	}
+	return (handle_general_word(lx));
+}
+
+t_lxstate	handle_general(t_lexer *lx)
+{
+	if (ft_isspace(*lx->cur))
+	{
+		emit_tok(lx);
+		while (ft_isspace(*lx->cur))
+			lx_advance(lx);
+	}
+	if (*lx->cur == '\0')
+		return (LX_EOF);
+	if (*lx->cur == '|' && lx->cur[1] == '|')
+		return (emit_op(lx, T_OR_IF, 2));
+	if (*lx->cur == '&' && lx->cur[1] == '&')
+		return (emit_op(lx, T_AND_IF, 2));
+	if (*lx->cur == '<' && lx->cur[1] == '<')
+		return (emit_op(lx, T_HEREDOC, 2));
+	if (*lx->cur == '>' && lx->cur[1] == '>')
+		return (emit_op(lx, T_APPEND, 2));
+	if (*lx->cur == '|')
+		return (emit_op(lx, T_PIPE, 1));
+	if (*lx->cur == '<')
+		return (emit_op(lx, T_INFILE, 1));
+	if (*lx->cur == '>')
+		return (emit_op(lx, T_OUTFILE, 1));
+	if (*lx->cur == '(')
+		return (emit_op(lx, T_LPAREN, 1));
+	return (handle_general_2(lx));
+}
